Fixed import error and fahrkostenpauschle step precondition
