+++ webapp/app/forms/steps/lotse/pauschbetrag.py
@@ -19,6 +19,7 @@ from app.forms.steps.lotse.merkzeichen import StepMerkzeichenPersonA, StepMerkze
     HasMerkzeichenPersonAPrecondition, HasMerkzeichenPersonBPrecondition
 from app.forms.steps.lotse.utils import get_number_of_users
 from app.forms.steps.lotse.has_disability import HasDisabilityPersonAPrecondition, HasDisabilityPersonBPrecondition
+from app.model.disability_data import DisabilityModel
 
 
 def calculate_pauschbetrag(has_pflegegrad=None, disability_degree=None, has_merkzeichen_bl=False, has_merkzeichen_tbl=False, has_merkzeichen_h=False):
@@ -57,37 +58,6 @@ def calculate_pauschbetrag(has_pflegegrad=None, disability_degree=None, has_merk
 
     return 0
 
-
-class DisabilityModel(BaseModel):
-    person_a_has_disability: Optional[str]
-    person_a_has_pflegegrad: Optional[str]
-    person_a_disability_degree: Optional[int]
-    person_a_has_merkzeichen_bl: Optional[bool]
-    person_a_has_merkzeichen_tbl: Optional[bool]
-    person_a_has_merkzeichen_g: Optional[bool]
-    person_a_has_merkzeichen_ag: Optional[bool]
-    person_a_has_merkzeichen_h: Optional[bool]
-
-    person_b_has_disability: Optional[str]
-    person_b_has_pflegegrad: Optional[str]
-    person_b_disability_degree: Optional[int]
-    person_b_has_merkzeichen_bl: Optional[bool]
-    person_b_has_merkzeichen_tbl: Optional[bool]
-    person_b_has_merkzeichen_g: Optional[bool]
-    person_b_has_merkzeichen_ag: Optional[bool]
-    person_b_has_merkzeichen_h: Optional[bool]
-
-    @root_validator()
-    def convert_disability_degree_to_int(cls, values):
-        values['person_a_disability_degree'] = int(values['person_a_disability_degree']) \
-            if values['person_a_disability_degree'] \
-            else None
-        values['person_b_disability_degree'] = int(values['person_b_disability_degree']) \
-            if values['person_b_disability_degree'] \
-            else None
-        return values
-
-
 class HasPauschbetragClaimPersonAPrecondition(DisabilityModel):
     _step_to_redirect_to = StepMerkzeichenPersonA.name
     _message_to_flash = _l('form.lotse.skip_reason.has_no_pauschbetrag_claim')
@@ -122,48 +92,6 @@ class HasPauschbetragClaimPersonBPrecondition(DisabilityModel):
         )
 
         if pauschbetrag_claim == 0:
-            raise ValidationError
-        return values
-
-
-class HasFahrkostenpauschaleClaimPersonAPrecondition(DisabilityModel):
-    _step_to_redirect_to = StepMerkzeichenPersonA.name
-    _message_to_flash = _l('form.lotse.skip_reason.has_fahrkostenpauschale_claim')
-
-    @root_validator(skip_on_failure=True)
-    def has_to_have_fahrkostenpauschale_not_0(cls, values):
-        fahrkostenpauschale_claim = calculate_fahrkostenpauschale(
-            has_pflegegrad=values.get('person_a_has_pflegegrad', None),
-            disability_degree=values.get('person_a_disability_degree', None),
-            has_merkzeichen_bl=values.get('person_a_has_merkzeichen_bl', False),
-            has_merkzeichen_tbl=values.get('person_a_has_merkzeichen_bl', False),
-            has_merkzeichen_h=values.get('person_a_has_merkzeichen_bl', False),
-            has_merkzeichen_ag=values.get('person_a_has_merkzeichen_tbl', False),
-            has_merkzeichen_g=values.get('person_a_has_merkzeichen_h', False)
-        )
-
-        if fahrkostenpauschale_claim == 0:
-            raise ValidationError
-        return values
-
-
-class HasFahrkostenpauschaleClaimPersonBPrecondition(DisabilityModel):
-    _step_to_redirect_to = StepMerkzeichenPersonB.name
-    _message_to_flash = _l('form.lotse.skip_reason.has_fahrkostenpauschale_claim')
-
-    @root_validator(skip_on_failure=True)
-    def has_to_have_fahrkostenpauschale_not_0(cls, values):
-        fahrkostenpauschale_claim = calculate_fahrkostenpauschale(
-            has_pflegegrad=values.get('person_abhas_pflegegrad', None),
-            disability_degree=values.get('person_abdisability_degree', None),
-            has_merkzeichen_bl=values.get('person_abhas_merkzeichen_bl', False),
-            has_merkzeichen_tbl=values.get('person_abhas_merkzeichen_bl', False),
-            has_merkzeichen_h=values.get('person_abhas_merkzeichen_bl', False),
-            has_merkzeichen_ag=values.get('person_abhas_merkzeichen_tbl', False),
-            has_merkzeichen_g=values.get('person_abhas_merkzeichen_h', False)
-        )
-
-        if fahrkostenpauschale_claim == 0:
             raise ValidationError
         return values
